chore: put back other services to main
