+++ apps/ui/.dockerignore
@@ -1,0 +1,5 @@
+.svelte-kit
+k8s
+build
+node_modules
+tests
