+++ apps/ui/k8s/templates/tests/test-connection.yaml
@@ -1,0 +1,15 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  name: "{{ include "ui.fullname" . }}-test-connection"
+  labels:
+    {{- include "ui.labels" . | nindent 4 }}
+  annotations:
+    "helm.sh/hook": test
+spec:
+  containers:
+    - name: wget
+      image: busybox
+      command: ['wget']
+      args: ['{{ include "ui.fullname" . }}:{{ .Values.service.port }}']
+  restartPolicy: Never
