--- apps/ui/src/changelog.md
@@ -1,2 +1,0 @@
-- v0.2.0
-- v0.1.0 - list all available bills in timeline view
