+++ apps/ui/src/routes/+layout.svelte
@@ -1,0 +1,18 @@
+<slot />
+
+<style>
+  @import url("https://fonts.googleapis.com/css2?family=Work+Sans:wght@200;400;900&display=swap");
+  @import url("https://fonts.googleapis.com/css2?family=Cutive+Mono&display=swap");
+
+  :global(body, html) {
+    margin: 0;
+    padding: 0;
+  }
+  :global(*) {
+    font-family: Quicksand, "Work Sans", sans-serif;
+  }
+
+  :global(body){
+    height: 100vh;
+  }
+</style>
