+++ apps/ui/src/routes/+page.server.ts
@@ -1,0 +1,13 @@
+/** @type {import('./$types').PageLoad} */
+export async function load() {
+    const appInfoResponse = await fetch(process.env.BILLS_SERVICE || 'http://bills')
+    const appInfo = await appInfoResponse.json()
+
+    const billsResponse = await fetch(process.env.BILLS_SERVICE || 'http://bills')
+    const bills = await billsResponse.json()
+    
+    return {
+        appInfo,
+        bills
+    };
+}
