--- apps/ui/src/src/lib/stores/urql.ts
@@ -1,8 +1,0 @@
-import { createClient, setContextClient } from "@urql/svelte";
-import { readable } from "svelte/store";
-
-const client = createClient({
-  url: `/graphql`,
-});
-
-export const urql = readable(client)
