--- apps/ui/src/src/routes/about/+page.svelte
@@ -1,1 +1,0 @@
-lel
