--- apps/ui/src/vite.config.ts
@@ -1,17 +1,0 @@
-import { sveltekit } from '@sveltejs/kit/vite';
-import type { UserConfig } from 'vite';
-
-const config: UserConfig = {
-	plugins: [sveltekit()],
-	optimizeDeps: {
-		exclude: ['@urql/svelte'],
-	},
-	server: {
-		proxy: {
-			// https://stackoverflow.com/questions/64677212/how-to-configure-proxy-in-vite
-			'/graphql': `${process.env.BILLS_SERVICE}/graphql`
-		}
-	}
-};
-
-export default config;
